# Add RVV Optimizations for RISC-V Architecture Platforms