--- a/tests/bench/Makefile
+++ b/tests/bench/Makefile
@@ -52,11 +52,14 @@ benchHash_avx2: CXXFLAGS += -mavx2
 benchHash_avx512: CFLAGS   += -mavx512f
 benchHash_avx512: CXXFLAGS += -mavx512f
 
+benchHash_rvv: CFLAGS   += -march=rv64gcv -O3
+benchHash_rvv: CXXFLAGS += -march=rv64gcv -O3
+
 benchHash_hw: CPPFLAGS += -DHARDWARE_SUPPORT
 benchHash_hw: CFLAGS   += -mavx2 -maes
 benchHash_hw: CXXFLAGS += -mavx2 -mpclmul -std=c++14
 
-benchHash benchHash32 benchHash_avx2 benchHash_avx512 benchHash_nosimd benchHash_hw: $(OBJ_LIST)
+benchHash benchHash32 benchHash_avx2 benchHash_avx512 benchHash_nosimd benchHash_hw benchHash_rvv: $(OBJ_LIST)
 	$(CXX) $(CPPFLAGS) $(CXXFLAGS) $^ $(LDFLAGS) -o $@
 
 
@@ -68,4 +71,4 @@ benchHash.o: benchHash.h
 
 
 clean:
-	$(RM) *.o benchHash benchHash32 benchHash_avx2 benchHash_avx512 benchHash_hw
+	$(RM) *.o benchHash benchHash32 benchHash_avx2 benchHash_avx512 benchHash_hw benchHash_rvv