@Cyan4973 2025-06-13
> by the way,when I use `CFLAGS="-march=rv64gcv -O3 -DXXH_FORCE_MEMORY_ACCESS=0" make check`, Performance with -O3 optimization is unexpectedly lower than with -O2.

This can indeed happen.

Generally, the main reason is that `-O3` tends to bloat the binary size, by generating too many specialized code paths. In a micro benchmark always hitting the same code paths, it's probably a win. But in a real world performance, other factors come in. Larger binary size also means instruction cache get filled more quickly. It can evict other tasks from there. Jumps are larger. etc. The impact on instruction cache, which is often quite small, is likely the larger culprit.

If `-O2` and `-O3` give roughly the same performance, I would recommend `-O2`.

Problem is, it's difficult to reach this conclusion _for all targets simultaneously_. It's likely true for some, but not others, that still benefit from `-O3`. Hence why the default is generally `-O3`, because it's supposed to generate the fastest binaries, and this is generally true, just not always.