@zijianli1234 2025-06-14
I understand now, thank you so much.I noticed that in the AVX2-optimized code path, the optimization level is explicitly set to -O2 using GCC pragmas:
```
#if XXH_VECTOR == XXH_AVX2 /* AVX2 */ \
  && defined(__GNUC__) && !defined(__clang__) /* GCC, not Clang */ \
  && defined(__OPTIMIZE__) && XXH_SIZE_OPT <= 0 /* respect -O0 and -Os */
#  pragma GCC push_options
#  pragma GCC optimize("-O2")
#endif
```
Should I also use this approach in my code to resolve the problem?