@Cyan4973 2025-06-14
If you imply that `-O2` is always better than `-O3` for your `RVV` code path,
then yes.

This conclusion could be more difficult to reach for RISC-V architecture than for `AVX2` ones, due to the wider variety of chip designs. I personally don't know the domain enough to tell if it's always true, or generally true, or just sometimes true.